Add button to order form

+++ src/main/java/com/inlimited/springmvc/repository/api/IReceiptDAO.java
@@ -2,13 +2,12 @@ package com.inlimited.springmvc.repository.api;
 
 
 import com.inlimited.springmvc.entity.Receipt;
-import com.inlimited.springmvc.entity.User;
 
 import java.util.List;
 
 public interface IReceiptDAO {
     Receipt saveReceipt(Receipt receipt);
-    Receipt findReceiptById(int receipt_id);
+    Receipt findReceiptById(Long receipt_id);
     List<Receipt> findAllReceipt();
     void updateReceipt(Receipt receipt);
     void deleteReceipt(Integer receiptId);
